+++ NEWS.md
@@ -1,3 +1,17 @@
+# concurve 1.08
+
+## Major changes
+
+* Can produce consonance and surprisal functions for correlations via the `corrintervals` function.
+* Now able to construct consonance and surprisal functions from the point estiate, and confidence limits via the `reveng` function.
+* Graphs produced via the `plotpint` or `plotsint` function now able to take custom titles, subtitles, x-axis titles, and captions.
+
+# concurve 1.07
+
+## Major changes
+
+* Can now produce consonance and surprisal functions for survival data produced with the `survival` package.
+
 # concurve 1.06
 
 ## Major changes